requirements.txt
tqdm==4.62.3
trimesh
pyglet<2
scipy

pymeshlab
torch
numpy
libigl
pyvista

